Cambios varios en la funcionalidad de reportes del administrador

+++ src/main/java/Utils/ConsumoTransformador.java
@@ -11,12 +11,12 @@ public class ConsumoTransformador {
 	protected int id;
 	protected double longitud;
 	protected double latitud;
-	private float consumoTotal;
+	private double consumoTotal;
 	
 	
 	private int identificadorSistema;
 
-	public ConsumoTransformador(int unId, double unaLongitud, double unaLatitud, float unConsumoTotal) {
+	public ConsumoTransformador(int unId, double unaLongitud, double unaLatitud, double unConsumoTotal) {
 		
 		this.id = unId;
 		this.longitud = unaLongitud;
